Change Cell Parameters casting

See merge request compsci308_2021spring/cellsociety_team06!81

+++ src/cell_society/backend/simulation_initializer/CellParameters.java
@@ -39,7 +39,7 @@ public class CellParameters {
    * @return The mapped value as a integer
    */
   public Integer getAsInt(String key) {
-    return Integer.parseInt(SIMULATION_MAP.get(key));
+    return (int) Double.parseDouble(SIMULATION_MAP.get(key));
   }
 
   /**
